+++ build.gradle
@@ -27,7 +27,6 @@ subprojects {
     apply plugin: 'kotlin-spring'
     apply plugin: 'eclipse'
     apply plugin: 'io.spring.dependency-management'
-    apply plugin: "com.arenagod.gradle.MybatisGenerator"
 
     compileKotlin {
         kotlinOptions {
@@ -48,42 +47,46 @@ subprojects {
     }
 
     dependencies {
+        implementation("org.springframework:spring-web")
+        implementation('org.springframework.boot:spring-boot-starter-security')
+        implementation('com.fasterxml.jackson.module:jackson-module-kotlin')
         implementation('org.mybatis.spring.boot:mybatis-spring-boot-starter:1.3.2')
         implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
         implementation("org.jetbrains.kotlin:kotlin-reflect")
+
         runtimeOnly('mysql:mysql-connector-java')
+
+        testImplementation('org.springframework.security:spring-security-test')
     }
+
+    dependencyManagement {
+        imports { mavenBom("org.springframework.boot:spring-boot-dependencies:${springBootVersion}") }
+    }
+
+}
+
+project(':common') {
+    apply plugin: "com.arenagod.gradle.MybatisGenerator"
 
     configurations {
         mybatisGenerator
     }
-
+    // 「./gradlew mbGenerator」コマンドで作成
     mybatisGenerator {
         verbose = true
-        configFile = "${projectDir}/mybatis/generatorConfig.xml"
+        configFile = "${rootDir}/db/mybatis/generatorConfig.xml"
     }
 }
 
-project(':common') {
-    dependencies {
-        implementation("org.springframework:spring-web")
-        implementation('com.fasterxml.jackson.module:jackson-module-kotlin')
-    }
-    dependencyManagement {
-        imports { mavenBom("org.springframework.boot:spring-boot-dependencies:${springBootVersion}") }
-    }
-}
-
-project(':webapp') {
+project(':connect-web') {
     apply plugin: 'application'
     apply plugin: 'org.springframework.boot'
 
     dependencies {
-        implementation('org.springframework.boot:spring-boot-starter-security')
+        compile project(':common')
         implementation('org.springframework.boot:spring-boot-starter-thymeleaf')
         implementation('org.springframework.boot:spring-boot-starter-web')
         testImplementation('org.springframework.boot:spring-boot-starter-test')
-        testImplementation('org.springframework.security:spring-security-test')
     }
 
     mainClassName = 'com.nexti4hack.webapp.WebappApplication'
@@ -94,16 +97,12 @@ project(':webapp') {
     }
 }
 
-project(':api') {
+project(':connect-api') {
     apply plugin: 'application'
     apply plugin: 'org.springframework.boot'
 
     dependencies {
-        implementation("org.springframework:spring-web")
-        implementation('com.fasterxml.jackson.module:jackson-module-kotlin')
-    }
-    dependencyManagement {
-        imports { mavenBom("org.springframework.boot:spring-boot-dependencies:${springBootVersion}") }
+        compile project(':common')
     }
 
     mainClassName = 'com.nexti4hack.api.ApiApplication'
