+++ common/src/main/kotlin/com/nexti4hack/common/component/PasswordEncoder.kt
@@ -1,0 +1,27 @@
+package com.nexti4hack.common.component
+
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
+import org.springframework.stereotype.Component
+
+/**
+ * パスワード暗号化/照合コンポーネント
+ */
+@Component
+open class PasswordEncoder {
+
+    /**
+     * パスワード暗号化
+     */
+    fun encode(value: String):String {
+        val bCryptPasswordEncoder = BCryptPasswordEncoder()
+        return bCryptPasswordEncoder.encode(value)
+    }
+
+    /**
+     * パスワード照合
+     */
+    fun matches(rawPassword: String, encodedPassword: String): Boolean {
+        val bCryptPasswordEncoder = BCryptPasswordEncoder()
+        return bCryptPasswordEncoder.matches(rawPassword, encodedPassword)
+    }
+}
