+++ common/src/main/kotlin/com/nexti4hack/common/custom/mapper/CustomFunctionMapper.kt
@@ -1,0 +1,13 @@
+package com.nexti4hack.common.custom.mapper
+
+import org.apache.ibatis.annotations.Mapper
+
+/**
+ * ファンクション呼び出し
+ */
+@Mapper
+interface CustomFunctionMapper {
+
+    // key取得
+    fun selectKeyValue(): String
+}
