+++ connect-web/src/main/kotlin/com/nexti4hack/webapp/WebappApplication.kt
@@ -1,0 +1,14 @@
+package com.nexti4hack.webapp
+
+import org.mybatis.spring.annotation.MapperScan
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@MapperScan("com.nexti4hack")
+@SpringBootApplication(scanBasePackages=["com.nexti4hack"])
+class WebappApplication
+
+fun main(args: Array<String>) {
+	runApplication<WebappApplication>(*args)
+}
+
