+++ connect-web/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+# Application
+server.servlet.context-path=/webapp
+
+# MySQL
+spring.datasource.url=jdbc:mysql://127.0.0.1:13306/webapp
+spring.datasource.username=webapp
+spring.datasource.password=webapp
+spring.datasource.driverClassName=com.mysql.jdbc.Driver
