+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'nexti4hack-system'
-include 'common', 'webapp', 'api'
+include 'common', 'connect-web', 'connect-api'
