--- webapp/src/main/java/com/nexti4hack/db/mapper/UsersMapper.java
@@ -1,120 +1,0 @@
-package com.nexti4hack.db.mapper;
-
-import com.nexti4hack.db.model.Users;
-import com.nexti4hack.db.model.UsersExample;
-import java.util.List;
-import org.apache.ibatis.annotations.Param;
-
-public interface UsersMapper {
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    long countByExample(UsersExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int deleteByExample(UsersExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int deleteByPrimaryKey(String userId);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int insert(Users record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int insertSelective(Users record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    List<Users> selectByExampleWithBLOBs(UsersExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    List<Users> selectByExample(UsersExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    Users selectByPrimaryKey(String userId);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int updateByExampleSelective(@Param("record") Users record, @Param("example") UsersExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int updateByExampleWithBLOBs(@Param("record") Users record, @Param("example") UsersExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int updateByExample(@Param("record") Users record, @Param("example") UsersExample example);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKeySelective(Users record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKeyWithBLOBs(Users record);
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    int updateByPrimaryKey(Users record);
-}
