--- webapp/src/main/java/com/nexti4hack/db/model/Users.java
@@ -1,245 +1,0 @@
-package com.nexti4hack.db.model;
-
-import java.io.Serializable;
-import java.util.Date;
-
-public class Users implements Serializable {
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column users.user_id
-     *
-     * @mbg.generated
-     */
-    private String userId;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column users.password
-     *
-     * @mbg.generated
-     */
-    private String password;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column users.create_date
-     *
-     * @mbg.generated
-     */
-    private Date createDate;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column users.update_date
-     *
-     * @mbg.generated
-     */
-    private Date updateDate;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column users.create_user
-     *
-     * @mbg.generated
-     */
-    private String createUser;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column users.update_user
-     *
-     * @mbg.generated
-     */
-    private String updateUser;
-
-    /**
-     *
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database column users.login_id
-     *
-     * @mbg.generated
-     */
-    private byte[] loginId;
-
-    /**
-     * This field was generated by MyBatis Generator.
-     * This field corresponds to the database table users
-     *
-     * @mbg.generated
-     */
-    private static final long serialVersionUID = 1L;
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column users.user_id
-     *
-     * @return the value of users.user_id
-     *
-     * @mbg.generated
-     */
-    public String getUserId() {
-        return userId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column users.user_id
-     *
-     * @param userId the value for users.user_id
-     *
-     * @mbg.generated
-     */
-    public void setUserId(String userId) {
-        this.userId = userId == null ? null : userId.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column users.password
-     *
-     * @return the value of users.password
-     *
-     * @mbg.generated
-     */
-    public String getPassword() {
-        return password;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column users.password
-     *
-     * @param password the value for users.password
-     *
-     * @mbg.generated
-     */
-    public void setPassword(String password) {
-        this.password = password == null ? null : password.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column users.create_date
-     *
-     * @return the value of users.create_date
-     *
-     * @mbg.generated
-     */
-    public Date getCreateDate() {
-        return createDate;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column users.create_date
-     *
-     * @param createDate the value for users.create_date
-     *
-     * @mbg.generated
-     */
-    public void setCreateDate(Date createDate) {
-        this.createDate = createDate;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column users.update_date
-     *
-     * @return the value of users.update_date
-     *
-     * @mbg.generated
-     */
-    public Date getUpdateDate() {
-        return updateDate;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column users.update_date
-     *
-     * @param updateDate the value for users.update_date
-     *
-     * @mbg.generated
-     */
-    public void setUpdateDate(Date updateDate) {
-        this.updateDate = updateDate;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column users.create_user
-     *
-     * @return the value of users.create_user
-     *
-     * @mbg.generated
-     */
-    public String getCreateUser() {
-        return createUser;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column users.create_user
-     *
-     * @param createUser the value for users.create_user
-     *
-     * @mbg.generated
-     */
-    public void setCreateUser(String createUser) {
-        this.createUser = createUser == null ? null : createUser.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column users.update_user
-     *
-     * @return the value of users.update_user
-     *
-     * @mbg.generated
-     */
-    public String getUpdateUser() {
-        return updateUser;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column users.update_user
-     *
-     * @param updateUser the value for users.update_user
-     *
-     * @mbg.generated
-     */
-    public void setUpdateUser(String updateUser) {
-        this.updateUser = updateUser == null ? null : updateUser.trim();
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method returns the value of the database column users.login_id
-     *
-     * @return the value of users.login_id
-     *
-     * @mbg.generated
-     */
-    public byte[] getLoginId() {
-        return loginId;
-    }
-
-    /**
-     * This method was generated by MyBatis Generator.
-     * This method sets the value of the database column users.login_id
-     *
-     * @param loginId the value for users.login_id
-     *
-     * @mbg.generated
-     */
-    public void setLoginId(byte[] loginId) {
-        this.loginId = loginId;
-    }
-}
