--- webapp/src/main/kotlin/com/nexti4hack/webapp/WebappApplication.kt
@@ -1,14 +1,0 @@
-package com.nexti4hack.webapp
-
-import org.mybatis.spring.annotation.MapperScan
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@MapperScan("com.nexti4hack.db.mapper,com.nexti4hack.webapp.mapper")
-@SpringBootApplication
-class WebappApplication
-
-fun main(args: Array<String>) {
-	runApplication<WebappApplication>(*args)
-}
-
