--- webapp/src/main/resources/application.properties
@@ -1,8 +1,0 @@
-# Application
-server.servlet.context-path=/webapp
-
-# MySQL
-spring.datasource.url=jdbc:mysql://127.0.0.1:3306/webapp
-spring.datasource.username=webapp
-spring.datasource.password=webapp
-spring.datasource.driverClassName=com.mysql.jdbc.Driver
