--- webapp/src/main/resources/com/nexti4hack/webapp/mapper/CustomFunctionMapper.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.nexti4hack.webapp.mapper.CustomFunctionMapper">
-    <select id="selectKeyValue" resultType="java.lang.String" >
-        select get_key() from dual
-    </select>
-</mapper>
